osipub/requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-spanner==1.7.1
googleapis-common-protos==1.5.6
grpc-google-iam-v1==0.11.4
grpcio==1.18.0
grpcio-gcp==0.2.2
idna==2.8
pkg-resources==0.0.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
websocket-client==0.54.0
